+++ central/pod/store/postgres/store.go
@@ -95,7 +95,7 @@ func createTablePodsLiveInstances(ctx context.Context, db *pgxpool.Pool) {
 	table := `
 create table if not exists pods_LiveInstances (
     pods_Id varchar,
-    idx numeric,
+    idx integer,
     InstanceId_ContainerRuntime integer,
     InstanceId_Id varchar,
     InstanceId_Node varchar,
@@ -105,7 +105,7 @@ create table if not exists pods_LiveInstances (
     Started timestamp,
     ImageDigest varchar,
     Finished timestamp,
-    ExitCode numeric,
+    ExitCode integer,
     TerminationReason varchar,
     PRIMARY KEY(pods_Id, idx),
     CONSTRAINT fk_parent_table FOREIGN KEY (pods_Id) REFERENCES pods(Id) ON DELETE CASCADE
@@ -133,7 +133,7 @@ func createTablePodsTerminatedInstances(ctx context.Context, db *pgxpool.Pool) {
 	table := `
 create table if not exists pods_TerminatedInstances (
     pods_Id varchar,
-    idx numeric,
+    idx integer,
     PRIMARY KEY(pods_Id, idx),
     CONSTRAINT fk_parent_table FOREIGN KEY (pods_Id) REFERENCES pods(Id) ON DELETE CASCADE
 )
@@ -161,8 +161,8 @@ func createTablePodsTerminatedInstancesInstances(ctx context.Context, db *pgxpoo
 	table := `
 create table if not exists pods_TerminatedInstances_Instances (
     pods_Id varchar,
-    pods_TerminatedInstances_idx numeric,
-    idx numeric,
+    pods_TerminatedInstances_idx integer,
+    idx integer,
     InstanceId_ContainerRuntime integer,
     InstanceId_Id varchar,
     InstanceId_Node varchar,
@@ -172,7 +172,7 @@ create table if not exists pods_TerminatedInstances_Instances (
     Started timestamp,
     ImageDigest varchar,
     Finished timestamp,
-    ExitCode numeric,
+    ExitCode integer,
     TerminationReason varchar,
     PRIMARY KEY(pods_Id, pods_TerminatedInstances_idx, idx),
     CONSTRAINT fk_parent_table FOREIGN KEY (pods_Id, pods_TerminatedInstances_idx) REFERENCES pods_TerminatedInstances(pods_Id, idx) ON DELETE CASCADE
